% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.addnode}
\alias{anfis.addnode}
\title{ANFIS node adder}
\usage{
anfis.addnode(anfis, idx, value = NULL, dedo = NULL, fan.in = NULL,
  fan.out = NULL, mf.type = NULL, mf.params = NULL, range = NULL,
  mf.name = NULL)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{idx}{The index of the layer where node is going to be added.}

\item{value}{node value}

\item{dedo}{The derivative of the error to the output of this node.}

\item{fan.in}{The indexes of input nodes from previous layer.}

\item{fan.out}{The indexes of output nodes in next layer.}

\item{mf.type}{membership function type, if this is a node of fuzzy membership function.}

\item{mf.params}{The parameters for the membership function}

\item{range}{The universe of discourse for the membership function}

\item{mf.name}{The name of membership function}
}
\value{
The ANFIS model with added node.
}
\description{
To add a node to the specified layer of a given ANFIS model.
}
\details{
This function is not designed for external use.
}
\author{
Chao Chen
}

