% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.builder}
\alias{anfis.builder}
\title{ANFIS model builder}
\usage{
anfis.builder(fis)
}
\arguments{
\item{fis}{A fuzzy inference system model initialised by \code{\link{newfis}}.}
}
\value{
An ANFIS model
}
\description{
To build an ANFIS model from an existing FIS model
}
\examples{
fis <- anfis.tipper()
anfis <- anfis.builder(fis)
}
\author{
Chao Chen
}
\references{
An extended ANFIS architecture and its learning properties for type-1 and interval type-2 models \url{https://doi.org/10.1109/FUZZ-IEEE.2016.7737742}
}

