% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dO1}
\alias{anfis.dE.dO1}
\title{anfis.dE.dO1}
\usage{
anfis.dE.dO1(anfis, output.L1, de.do2, do2.do1)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L1}{The output of nodes in Layer 1}

\item{de.do2}{The derivatives of output error with respect to output.L2}

\item{do2.do1}{The derivatives of output.L2 with respect to output.L1.}
}
\value{
The derivatives of output error with respect to output.L1.
}
\description{
to calculate the derivatives of output error with respect to output.L1.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
