% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dO2}
\alias{anfis.dE.dO2}
\title{anfis.dE.dO2}
\usage{
anfis.dE.dO2(de.do3, do3.do2)
}
\arguments{
\item{de.do3}{The derivatives of output error with respect to output.L3}

\item{do3.do2}{The derivatives of output.L3 with respect to output.L2.}
}
\value{
The derivatives of output error with respect to output.L2.
}
\description{
to calculate the derivatives of output error with respect to output.L2.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
