% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dO3}
\alias{anfis.dE.dO3}
\title{anfis.dE.dO3}
\usage{
anfis.dE.dO3(de.do4, do4.do3, output.L3)
}
\arguments{
\item{de.do4}{The derivatives of output error with respect to output.L4}

\item{do4.do3}{The derivatives of output.L4 with respect to output.L3.}

\item{output.L3}{The output of nodes in Layer 3.}
}
\value{
The derivatives of output error with respect to output.L3.
}
\description{
to calculate the derivatives of output error with respect to output.L3.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
