% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dO4}
\alias{anfis.dE.dO4}
\title{anfis.dE.dO4}
\usage{
anfis.dE.dO4(anfis, de.do5, do5.do4)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{de.do5}{The derivatives of output error with respect to output.L5}

\item{do5.do4}{The derivatives of output.L5 with respect to output.L4.}
}
\value{
The derivatives of output error with respect to output.L4.
}
\description{
to calculate the derivatives of output error with respect to output.L4.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
