% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dO5}
\alias{anfis.dE.dO5}
\title{anfis.dE.dO5}
\usage{
anfis.dE.dO5(output.L5, y)
}
\arguments{
\item{output.L5}{the model outputs}

\item{y}{the target outputs}
}
\value{
The derivatives of output error with respect to output.L5
}
\description{
To calculate the derivatives of output error with respect to output.L5.
NOTE: currently, only single output in L5 is supported
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
