% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dP1.it2gbellmf}
\alias{anfis.dE.dP1.it2gbellmf}
\title{anfis.dE.dP1.it2gbellmf}
\usage{
anfis.dE.dP1.it2gbellmf(de.do1, x, mf.params)
}
\arguments{
\item{de.do1}{The derivatives of output error with respect to output.L1}

\item{x}{The crisp input}

\item{mf.params}{parameters for membership functions}
}
\description{
to calculate the derivatives of E versus mf.params.L1 for it2gbellmf
NOTE: only singleton fuzzification is supported
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
