% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dE.dP4}
\alias{anfis.dE.dP4}
\title{anfis.dE.dP4}
\usage{
anfis.dE.dP4(anfis, de.do4, output.L3, input.stack)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{de.do4}{The derivatives of output error with respect to output.L4}

\item{output.L3}{The output of nodes in Layer 3}

\item{input.stack}{The input data pairs.}
}
\value{
The derivatives of output error with respect to parameters in Layer 4.
}
\description{
To calculate the derivatives of output error with respect to parameters in Layer 4.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
