% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dMF.dP.gbellmf}
\alias{anfis.dMF.dP.gbellmf}
\title{anfis.dMF.dP.gbellmf}
\usage{
anfis.dMF.dP.gbellmf(x, mf.params)
}
\arguments{
\item{x}{The crisp input}

\item{mf.params}{parameters for membership functions}
}
\description{
to calculate the derivatives of membership grades with respect to its parameters
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
