% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dO2.dO1}
\alias{anfis.dO2.dO1}
\title{anfis.dO2.dO1}
\usage{
anfis.dO2.dO1(anfis, output.L2, output.L1)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L2}{The output of nodes in Layer 2}

\item{output.L1}{The output of nodes in Layer 1}
}
\value{
The derivatives of output.L2 with respect to output.L1. do2[j].do1[i] <- do2.do1[[i]][[which(fan.out==j)]]
}
\description{
To calculate the derivatives of output.L2 with respect to output.L1.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
