% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dO3.dO2}
\alias{anfis.dO3.dO2}
\title{anfis.dO3.dO2}
\usage{
anfis.dO3.dO2(anfis, output.L2, output.L2.which)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L2}{The output of nodes in Layer 2}

\item{output.L2.which}{A list of matrix indicating which output (w.lower, w.upper) in layer 2 should be used by the ekm algorithm}
}
\value{
The derivatives of output.L3 with respect to output.L2. do3.left[j].do2[i] <- do3.do2[[i]][[1]][[j]]
}
\description{
To calculate the derivatives of output.L3 with respect to output.L2.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
