% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dO4.dO3}
\alias{anfis.dO4.dO3}
\title{anfis.dO4.dO3}
\usage{
anfis.dO4.dO3(output.L4, output.L4.mf)
}
\arguments{
\item{output.L4}{The output of nodes in Layer 4}

\item{output.L4.mf}{The membership grades of the membership functions of nodes in Layer 4}
}
\value{
The derivatives of output.L4 with respect to output.L3.
}
\description{
To calculate the derivatives of output.L4 with respect to output.L3.
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
