% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.dO5.dO4}
\alias{anfis.dO5.dO4}
\title{anfis.dO5.dO4}
\usage{
anfis.dO5.dO4(output.L4)
}
\arguments{
\item{output.L4}{The output of nodes in Layer 4.}
}
\value{
The derivatives of output.L5 with respect to output.L4.
}
\description{
To calculate the derivatives of output.L5 with respect to output.L4.
NOTE: currently, only single output in L5 is supported
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
}
\author{
Chao Chen
}
