% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.eval}
\alias{anfis.eval}
\title{ANFIS evaluator}
\usage{
anfis.eval(anfis, input.stack)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{input.stack}{The input data}
}
\value{
The output of the anfis for given input data.
}
\description{
To evaluate a ANFIS model with input data
}
\examples{
fis <- anfis.tipper()
anfis <- anfis.builder(fis)
data.num <- 5
input.num <- length(fis$input)
input.stack <- matrix(rnorm(data.num*input.num), ncol=input.num)
y <- matrix(rnorm(data.num))
data.trn <- cbind(input.stack, y)
anfis.eval(anfis, input.stack)
}
\author{
Chao Chen
}
