% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L1.eval}
\alias{anfis.L1.eval}
\title{The evaluator for nodes in Layer 1}
\usage{
anfis.L1.eval(anfis, output.LI, input.stack)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.LI}{The output of nodes in Layer I}

\item{input.stack}{The input data}
}
\value{
The output of nodes in Layer 1
}
\description{
To evaluate the antecedent layer (L1) of anfis
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
