% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L2.eval}
\alias{anfis.L2.eval}
\title{The evaluator for nodes in Layer 2}
\usage{
anfis.L2.eval(anfis, output.L1)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L1}{The output of nodes in Layer 1}
}
\value{
The output of nodes in Layer 2
}
\description{
To evaluate the nodes in Layer 2 of the given ANFIS model
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
