% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L2.which}
\alias{anfis.L2.which}
\title{L2.which}
\usage{
anfis.L2.which(anfis, output.L2, output.L4.mf)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L2}{The output of nodes in Layer 2}

\item{output.L4.mf}{The linear membership grades of nodes in Layer 4}
}
\value{
A list of matrix indicating which output (w.lower, w.upper) in layer 2 should be used by the ekm algorithm
}
\description{
To determin which output (w.lower, w.upper) to be used by the ekm algorithm
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
