% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L3.eval}
\alias{anfis.L3.eval}
\title{The evaluator for nodes in Layer 3}
\usage{
anfis.L3.eval(anfis, output.L2, output.L2.which)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L2}{The output of nodes in Layer 2}

\item{output.L2.which}{A list of matrix indicating which output (w.lower, w.upper) in layer 2 should be used by the ekm algorithm}
}
\value{
The output of nodes in Layer 3
}
\description{
To evaluate the nodes in Layer 3 of the given ANFIS model
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
