% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L4.eval}
\alias{anfis.L4.eval}
\title{The evaluator for nodes in Layer 4}
\usage{
anfis.L4.eval(output.L3, output.L4.mf)
}
\arguments{
\item{output.L3}{The output of nodes in Layer 3}

\item{output.L4.mf}{The membership grades of the membership functions of nodes in Layer 4}
}
\value{
The output of nodes in Layer 4
}
\description{
To evaluate the nodes in Layer 4
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
