% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L4.mf.eval}
\alias{anfis.L4.mf.eval}
\title{The evaluator for membership functions of nodes in Layer 1}
\usage{
anfis.L4.mf.eval(anfis, input.stack)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{input.stack}{The input data}
}
\value{
The membership grades of the membership functions of nodes in Layer 4
}
\description{
To evaluate the membership functions of nodes in Layer 4
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
