% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.L5.eval}
\alias{anfis.L5.eval}
\title{The evaluator for nodes in Layer 5}
\usage{
anfis.L5.eval(output.L4)
}
\arguments{
\item{output.L4}{The output of nodes in Layer 4}
}
\value{
The output of nodes in Layer 5
}
\description{
To evaluate the nodes in Layer 5
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
