% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.LI.eval}
\alias{anfis.LI.eval}
\title{The evaluator for nodes in Layer I}
\usage{
anfis.LI.eval(anfis, input.stack)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{input.stack}{The input data}
}
\value{
The output of nodes in Layer I
}
\description{
To evaluate the input Layer (LI) of anfis
}
\details{
This function is not recommended for external use, but can be used for debugging or learning.
See the source code of \code{\link{anfis.eval}} for usage.
}
\author{
Chao Chen
}
