% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.optimise.gradient}
\alias{anfis.optimise.gradient}
\title{The gradient algorithm for ANFIS optimisation}
\usage{
anfis.optimise.gradient(anfis, L, de.dp, stepsize)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{L}{The index of the layer in which the parameters to be optimised}

\item{de.dp}{The derivatives of the output error to the parameters}

\item{stepsize}{the stepsize for updating the parameters}
}
\value{
The ANFIS model with updated parameters
}
\description{
To update the mf.params in Layer 1 or Layer 4 by the gradient decent methed
}
\details{
This function is not designed for external use.
}
\author{
Chao Chen
}

