% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANFIS.R
\name{anfis.optimise.lse}
\alias{anfis.optimise.lse}
\title{The LSE algorithm for ANFIS optimisation}
\usage{
anfis.optimise.lse(anfis, output.L3, input.stack, target, online = 0,
  lambda = 1)
}
\arguments{
\item{anfis}{The given ANFIS model}

\item{output.L3}{The output of nodes from Layer 3}

\item{input.stack}{The input data (training data)}

\item{target}{The output data (training data)}

\item{online}{0 -- batch; 1,2 -- online}

\item{lambda}{the forgetting rate}
}
\value{
The ANFIS model with updated parameters
}
\description{
to update the mf.params in L4 by the LSE algorithm
}
\details{
This function is not designed for external use.
}
\author{
Chao Chen
}

