% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{anfis.tipper}
\alias{anfis.tipper}
\title{Produces an example fis object which can be used for ANFIS.}
\usage{
anfis.tipper()
}
\value{
A fis is return
}
\description{
A function used primarily for example purposes, it creates a fis with two input (service & food), output variables (tip) and their membership functions.
}
\examples{
fis <- anfis.tipper()
}
