% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinality.R
\name{cardinality}
\alias{cardinality}
\title{Cardinality of a fuzzy number}
\usage{
cardinality(fx,
           h = 0.01)
}
\arguments{
\item{fx}{A fuzzy number given by a \code{fzn} object.}

\item{h}{Numerical, between 0 and 1. Step for alphas in "spline" interpolation.}
}
\value{
A number that approximates the value of the cardinality.
}
\description{
This function estimates/approximates the value of the cardinality (area) of a fuzzy number given by
a \code{fzn} object.
}
\examples{
fx <- fzn(alpha = c(0, 1),
          l = c(3, 4),
          u = c(7, 6))
plot(fx)
cardinality(fx)

}
