% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{defuzz}
\alias{defuzz}
\title{Defuzzify a set of values.}
\usage{
defuzz(x, mf, type)
}
\arguments{
\item{x}{The range to be applied in the function (numeric vector).}

\item{mf}{The values to be applied in the function (numeric vector).}

\item{type}{The defuzzification method type, which should be either 'centroid', 'bisector', 'mom', 'som' or 'lom'.}
}
\value{
Returns a defuzzified crisp value (double).
}
\description{
Defuzzifies a given set of values using a specified range and defuzzification type producing a crisp value.
}
\examples{
Crisp_value = defuzz(1:10, c(1.5, 5), "centroid")
}
