% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{-.fzn}
\alias{-.fzn}
\alias{-}
\title{Substraction of fuzzy numbers}
\usage{
fx1 - fx2
}
\arguments{
\item{fx1}{First fuzzy (or crisp) number, or fuzzy array.}

\item{fx2}{Second fuzzy (or crisp) number, or fuzzy array.}
}
\value{
The fuzzy substraction (fx1 - fx2) as an object of class \code{fzn} or \code{fzarray}.
}
\description{
Substracts two fuzzy numbers of class \code{fzn} or one fuzzy number to a crisp number (or viceversa).
Also works for fuzzy arrays of class \code{fzarray}.
}
\examples{
x1 <- fzn(alpha = c(0, 0.1, 0.5, 0.8, 1),
          l = c(1, 3, 3, 4.5, 5),
          u = c(12, 10, 9, 6.5, 6))
x2 <- fzn(alpha = c(0, 0.25, 1),
          l = c(1, 3, 5),
          u = c(12, 10, 6))
par(mfrow = c(2, 2))
plotfzn(x1, main = "x1")
plotfzn(x2, main = "x2")
plotfzn(x1 - x2, main = "x1 - x2")

}
