% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_kaoliu_fzn.R
\name{eff_kaoliu_fzn}
\alias{eff_kaoliu_fzn}
\title{Extracts efficiencies from Kao-Liu result}
\usage{
eff_kaoliu_fzn(x)
}
\arguments{
\item{x}{Any \code{dea_fuzzy} object from any Kao-Liu model.}
}
\value{
Returns a \code{fzarray} object.
}
\description{
Extracts efficiencies as \code{fzarray} from Kao-Liu result.
}
