% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{evalfis}
\alias{evalfis}
\title{Evaluate a Fuzzy Inference System (fis)}
\usage{
evalfis(input_stack, fis, time = 1, point_n = 101, draw = FALSE)
}
\arguments{
\item{input_stack}{A matrix representing the input stack, number of inputs (columns) by number of outputs (rows).}

\item{fis}{A fis must be provided.}

\item{time}{default 1}

\item{point_n}{number of discretised points, default 101}

\item{draw}{whether to draw, TRUE or FALSE}
}
\value{
Returns a matrix of evaluated values.
}
\description{
Returns an evaluated crisp value for a given fis structure.
}
\examples{
Input_data <- matrix((1:2),1,2)
fis <- tipper()
evalfis(Input_data, fis)
}
