% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{evalmftype}
\alias{evalmftype}
\title{Evaluate fuzzy membership function with membership function type and parameters}
\usage{
evalmftype(x, mf.type, mf.params)
}
\arguments{
\item{x}{A generic element of U, which is the universe of discourse for a fuzzy set}

\item{mf.type}{The member function type}

\item{mf.params}{The parameters for a member function}
}
\value{
Membership grade(s)
}
\description{
To obtain the corresponding membership grade(s) for crisp input(s) x
}
\examples{
evalmftype(5, mf.type=gbellmf, mf.params=c(1,2,3))
evalmftype(1:10, mf.type=gbellmf, mf.params=c(1,2,3))
}
\author{
Chao Chen
}
