% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLAnova1Way}
\alias{FLAnova1Way}
\title{One-Way Analysis of Variance}
\usage{
FLAnova1Way(table, response, variable, where_clause = "")
}
\arguments{
\item{table}{An object of class \code{FLTable}}

\item{response}{column name for the data values}

\item{variable}{column name for the categorical variable}

\item{where_clause}{condition to filter out data from the table}
}
\value{
A \code{data.frame} with the following columns:
\item{Df}{Degrees of Freedom}
\item{Sum Sq}{Squared sum from mean}
\item{Mean Sq}{Mean squared sum}
\item{F value}{Value of the F-Statistic}
\item{Pr(>F)}{P-Value}
There are two rows in the result, one for the variable term in the model,
plus one for \code{Residuals} if there are any.

If all the values in all the categories are the same then the Sum Sq and
Mean Sq for \code{Residuals} are zero and the F Value and P Value are \code{NA}
}
\description{
Performs one-way analysis of variance(ANOVA).
}
\details{
One-way analysis of variance is used to test the difference between
two or more groups in a dataset based on a single attribute.
}
\examples{
\dontrun{

FLAnova1Way(table = tbl, response = "MPG", variable = "CarName")
# Select American brands only
FLAnova1Way(	table = tbl,
				response = "MPG",
				variable = "CarName",
				where_clause = "CarName IN ('Chrysler', 'Dodge', 'Ford')" )

}
}

