% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLAnova2Way.R
\name{FLAnova2Way}
\alias{FLAnova2Way}
\title{Two-Way Analysis of Variance}
\usage{
FLAnova2Way(table, response, variable1, variable2)
}
\arguments{
\item{table}{An object of class \code{FLTable}}

\item{response}{column name for the data values}

\item{variable1}{column name for the first categorical variable}

\item{variable2}{column name for the second categorical variable}
}
\value{
A \code{data.frame} with the following columns:
\item{Df}{Degrees of Freedom}
\item{Sum Sq}{Squared sum from mean}
\item{Mean Sq}{Mean squared sum}
\item{F value}{Value of the F-Statistic}
\item{Pr(>F)}{P-Value}
There are three rows in the result, one for each of the variable terms in the
model, plus one for \code{Residuals} if there are any.

If all the values in all the categories are the same then the Sum Sq and
Mean Sq for \code{Residuals} are zero and the F Value and P Value are \code{NA}
}
\description{
Performs two-way analysis of variance(ANOVA).
}
\details{
Two-way analysis of variance is used to test the difference between
two or more groups in a dataset based on two attribute.
}
\examples{
\dontrun{

FLAnova2Way(table = tbl, response = "MPG", variable1 = "CarName", variable2 = "Origin")

}
}

