% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLCholeskyDecomp}
\alias{FLCholeskyDecomp}
\title{Cholesky Decomposition of a matrix}
\usage{
FLCholeskyDecomp(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix.}
}
\value{
\code{FLCholeskyDecomp} returns an object of class FLMatrix mapped
to the lower triangular matrix generated by Cholesky decomposition, which
can be pulled to R using \code{FLFetchMatrix}.
}
\description{
Cholesky decomposition is a decomposition of a positive definite matrix
into the product of a lower triangular matrix and its conjugate transpose.
\code{FLCholeskyDecomp} computes the lower triangular matrix generated by
Cholesky decomposition.
}
\section{Constraints}{

Input can only be a positive definite matrix (n x n) with maximum dimension
limitations of (1000 x 1000).
}
\examples{
\dontrun{
library(RODBC)
connection      <- odbcConnect("Gandalf");
database        <- "FL_R_WRAP";
matrix_table    <- "tblMatrixMulti";
matrix_id_value <- 5;
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
out_matrix      <- FLCholeskyDecomp(in_matrix)
}
}

