% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLEigen.R
\name{FLEigen}
\alias{FLEigen}
\title{Eigen decomposition of a matrix}
\usage{
FLEigen(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix.}
}
\value{
\code{FLEigen} returns a list. The list has two components of class
\code{FLMatrix}: \code{vectors} and \code{values}. \code{"vectors"} is mapped
to a matrix containing the eigen vectors of the input matrix. \code{"values"}
is mapped to a diagonal matrix whose diagonal elements are the eigen values
of the input matrix.
}
\description{
Calculates the eigenvalues and eigenvectors of a square matrix (n x n).
}
\section{Constraints}{

Input can only be a square matrix (n x n) with maximum dimension limitations
of (700 x 700). The current FLEigen can only calculate real eigenvalues. If
the input matrix is not symmetric, it is suggested to verify the output
values.
}
\examples{
\dontrun{
connection      <- odbcConnect("Gandalf")
database        <- "FL_R_WRAP"
matrix_table    <- "tblMatrixMulti"
matrix_id_value <- 5
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
out_eigen       <- FLEigen(in_matrix)
}
}

