% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLMatrix.R
\name{FLFetchMatrix}
\alias{FLFetchMatrix}
\title{Pulls matrix into R}
\usage{
FLFetchMatrix(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix.}
}
\value{
\code{FLFetchMatrix} returns a data frame with three components which
are as follows.
\item{ROW_ID}{Row number of the matrix elements}
\item{COL_ID}{Column number of the matrix elements}
\item{CELL_VAL}{Value of the matrix elements}
}
\description{
\code{FLFetchMatrix} retrieves a matrix to which an FLMatrix object is
mapped to in Teradata.
}
\examples{
\dontrun{
connection      <- odbcConnect("Gandalf")
database        <- "FL_R_WRAP"
matrix_table    <- "tblMatrixMulti"
matrix_id_value <- 5
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
FLFetchMatrix(in_matrix)
}
}

