% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLFitDistr}
\alias{FLFitDistr}
\title{Distribution fitting}
\usage{
FLFitDistr(distribution_object, distribution, method = "MLE")
}
\arguments{
\item{distribution_object}{an object generated by \code{FLFitDistrObject}
which is mapped to a table in Teradata}

\item{distribution}{name of the distribution which is to be fitted. The
distribution to be fitted should be one from the list given in the
\code{Distributions} section}

\item{method}{indicator of the method used for estimating the fit distribution
parameters. Use \code{"MLE"} for Maximum Likelihood Estimation Method and
\code{"MDE"} for Minimum Distance Estimation Method}
}
\value{
\code{FLFitDistr} returns a \code{data.frame} which has the estimates
of the parameters along with the value of the log-likelihood or the distance.
The parameters depend on the distribution being fitted.
}
\description{
Fits a given distribution to a set of data points
}
\details{
\code{FLFitDistr} estimates parameter values of a given distribution
for a given data set. One of the methods used in fitting distribution is
by this function is Maximum Likelihood Estimation (MLE), that is estimating
parameters by maximizing log likelihood function. Apart from MLE, the function
can also estimate the parameters Minimum Distance Estimation (MDE), that is
estimating parameters by minimizing distance function.
}
\section{Note}{

\enumerate{
 \item When the starting values are very far away from the actual values or
 parameters, e.g. location, scale, shape, distance, probability, etc., do
 not converge, the functions will output NULLs.
 \item For certain distributions, if the input values contain zeros, the
 function may throw a system error such as, divide by zero or invalid
 arithmetic operation.
}
}

\section{Distributions}{

The following distributions can be fitted to a data set using \code{FLFitDistr}.
\enumerate{
 \item \code{Beta}
 \item \code{Binomial}
 \item \code{Cauchy}
 \item \code{ChiSq}
 \item \code{Normal}
 \item \code{Poisson}
 \item \code{Weibull}
}
}

