% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLLDA}
\alias{FLLDA}
\title{Linear Discriminant Analysis}
\usage{
FLLDA(table, primary_key, response, exclude = c(), class_spec = list(),
  where_clause = "", note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column of the table mapped to
\code{table}}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{note}
}
\value{
\code{FLLDA} returns an object of class \code{FLLDA}. The components
of this class mentioned below can be pulled in R using the generic
\code{FLFetch}.
\item{canonical_coeffs}{a \code{data.frame} which stores the the Canonical
Coefficients}
\item{fisher_coeffs}{a \code{data.frame} which stores the the Fisher
Coefficients}
\item{canonical_variates}{a \code{data.frame} which stores the Canonical
Variates}
\item{cross_tables}{a \code{data.frame} which stores the Cross Tables of
predicted and observed classfication}
}
\description{
\code{FLLDA} performs Fisher linear discriminant analysis (LDA).
}
\examples{
\dontrun{
connection <- odbcConnect("Gandalf")
db_name    <- FL_R_WRAP
table_name <- "tblirisdata"
# Create FLTable object
table      <-  FLTable(connection, db_name, table_name)
# Perform LDA
result     <- FLLDA( table,
                     primary_key = "ObsID",
                     response = "SpeciesID",
                     exclude = c("Species"))
# Fetch reults in R
ldaResult <- FLFetch(result)
}
}

