% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLLinRegr}
\alias{FLLinRegr}
\title{Linear Regression}
\usage{
FLLinRegr(table, primary_key, response, exclude = as.character(c()),
  class_spec = list(), where_clause = "",
  note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{note (a string)}
}
\value{
returns an object of class \code{FLLinRegr} whose components can be
pulled to R by running FLFetch. The class will then have 2 slots:

\code{stats}:
\item{RSquared}{R-Square}
\item{AdjRSquared}{Adjusted R-Square}
\item{StdErr}{Standard Error}
\item{DFRegression}{Regression Degree of Freedom}
\item{DFResidual}{Residual Degree of Freedom}
\item{DFTotal}{Total Degree of Freedom}
\item{SSRegression}{Regression Sum Squared}
\item{SSResidual}{Residual Sum Squared}
\item{SSTotal}{Total Sum Squared}
\item{MSRegression}{Regression Mean Sum Squared}
\item{MSResidual}{Residual Mean Sum Squared}
\item{FStat}{F-Statistic}
\item{SigFStat}{Significance of F-Statistic}
\item{ResidualAutoCorrel}{Lag-1 Auto Correlation between Residuals}
\item{WStat}{Durbin-Watson Test statistic}
\item{BPStat}{Breusch Pagan Test statistic}

\code{coeffs}:
\item{COEFFID}{Coefficient ID}
\item{VAR_TYPE}{Variable Type}
\item{COLUMN_NAME}{Variable Name}
\item{CATVALUE}{Category Name represented by dummy variable}
\item{CoeffValue}{They are coefficients, one for each explanatory variable, that represent the strength and type of relationship the explanatory variable has to the dependent variable.}
\item{StdErr}{Standard Error(an estimate of the standard deviation of the coefficient)}
\item{TStat}{t-Statistic(coefficient divided by its standard error.)}
\item{PValue}{P-Value of t-Statistic}
\item{NonZeroDensity}{Proportion of non-zero values for the explanatory variable}
\item{CorrelWithRes}{Correlation of the explanatory variable with the residuals}
}
\description{
Performs Linear Regression
}
\examples{
\dontrun{
}
}

