% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLLUDecomp}
\alias{FLLUDecomp}
\title{Lower-upper (lu) decomposition of a matrix}
\usage{
FLLUDecomp(matrix)
}
\arguments{
\item{matrix}{an object of class FLMatrix; \code{matrix} should be square.}
}
\value{
\code{FLLUDecomp} returns an object of class FLLUDecomp. An object
of class FLLUDecomp is a list containing three components of class FLMatrix
which are described as follows.
\item{l_matrix}{lower triangular matrix}
\item{u_matrix}{upper triangular matrix}
\item{perm_matrix}{permutation matrix}
All the three matrices of class FLLUDecomp can be pulled in R using
\code{FLFetchMatrix}.
}
\description{
The LU decomposition involves factorizing a matrix as the product of a lower
triangular matrix L and an upper triangular matrix U. Permutation matrix is
also provided in the output. If permutation matrix is not used in the decomposition, the output of
permutation matrix is an identity matrix.
}
\section{Constraints}{

Input matrix has a maximum dimension limitations of (1000 x 1000).
}
\examples{
\dontrun{
library(RODBC)
connection      <- odbcConnect("Gandalf");
database        <- "FL_R_WRAP";
matrix_table    <- "tblMatrixMulti";
matrix_id_value <- 5;
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
out_lu          <- FLLUDecomp(in_matrix)
}
}

