% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLMatchIt.R
\name{FLMatchIt}
\alias{FLMatchIt}
\title{Nearest neighbour matching}
\usage{
FLMatchIt(table, obs_id, treatment, prop_score, match_order)
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{obs_id}{name of the Observation ID column name}

\item{treatment}{name of the treatment column}

\item{prop_score}{name of the propensity score column}

\item{match_order}{name of the match order column}
}
\value{
\code{FLMatchtIt} returns an object of class \code{FLMatchIt}. An
object of this class is a list containing three components.
\item{odbc_connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}.}
\item{out_table_name}{name of the permanent table which stores the output
in \code{Teradata}. This table does not exist in the R workspace.}
\item{unmatched_obs_id}{a \code{data.frame} in R which stores the list of \code{obs_ids}
which are not matched. This data frame is pulled in R using the generic
\code{FLFetch}.}
}
\description{
\code{FLMatchIt} performs "nearest neighbor matching" without replacement.
The output of this function is a list of observation IDs which are not matched.
}
\examples{
\dontrun{
connection <- odbcConnect("Gandalf")
db_name    <- FL_R_WRAP
table_name <- "TblMatchItAlt"
# Create FLTable object
table      <- FLTable(connection, db_name, table_name)
# Perform nearest neighbour matching
res        <- FLMatchIt(table, "person_id", "exposure", "prob", "prob")
}
}

