% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLMatrix}
\alias{FLMatrix}
\title{Constructor function for matrix}
\usage{
FLMatrix(connection, database, matrix_table, matrix_id_value,
  matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
  cell_value = "Cell_Val")
}
\arguments{
\item{connection}{ODBC connection handle as returned by \code{\link[RODBC]{odbcConnect}}}

\item{database}{name of the database in \code{Teradata} which has the matrix
table}

\item{matrix_table}{name of the matrix table}

\item{matrix_id_value}{identifier for the input matrix}

\item{matrix_id,row_id,column_id,cell_value}{column names in \code{matrix_table}
which give the matrix_id_value, row number, column number and value of the
matrix element}
}
\value{
\code{FLMatrix} returns an object of class FLMatrix mapped
to a matrix, which can be pulled to R using \code{FLFetchMatrix}.
}
\description{
\code{FLMatrix} constructs an object of class \code{FLMatrix}. This object
     is used as the input for all the matrix operation functions.
}
\examples{
\dontrun{
connection      <- odbcConnect("Gandalf")
database        <- "FL_R_WRAP"
matrix_table    <- "tblMatrixMulti"
matrix_id_value <- 5
in_matrix       <- FLMatrix(connection, database, matrix_table, matrix_id_value,
                   matrix_id = "Matrix_ID", row_id = "Row_ID", column_id = "Col_ID",
                   cell_value = "Cell_Val")
}
}

