% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FLVIF}
\alias{FLVIF}
\title{Variance Inflation Factor}
\usage{
FLVIF(table, primary_key, response, exclude = c(), class_spec = list(),
  where_clause = "", note = "From RWrapper For DBLytix")
}
\arguments{
\item{table}{an object of class \code{FLTable}}

\item{primary_key}{name of primary key column of the table mapped to \code{table}}

\item{response}{name of the dependent variable column}

\item{exclude}{vector of names of the columns which are to be excluded}

\item{class_spec}{list that identifies the value of the categorical variable
which is to be used as reference when converting to dummy binary variables}

\item{where_clause}{condition to filter out data from the table}

\item{note}{free form string that will be stored with the results, typically
used to document the purpose of the analysis}
}
\value{
an object of class \code{FLVIF}
}
\description{
Performs variance inflation factor analysis on data
}
\details{
Variance Inflation Factor is used to identify redundant variables in
a dataset. The square root of the variance inflation factor tells you how
much larger the standard error is, compared with what it would be if that
variable were uncorrelated with the other predictor variables in the model.
}
\examples{
\dontrun{
tbl <- FLTable(DBConnect, "FL_R_WRAP", "tblAutoMpg")

Analysis <- FLVIF(	Tbl,
						primary_key = "ObsID",
						response = "MPG",
						exclude = c("CarNum","CarNumber"),
						class_spec = list(CarName = "Audi"))

Analysis <- FLFetch(Analysis)
}
}

