% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_ratio.R
\name{fuzz_ratio}
\alias{fuzz_ratio}
\title{Fuzzy Ratio}
\usage{
fuzz_ratio(x1, x2)
}
\arguments{
\item{x1}{a single string or vector of strings}

\item{x2}{a single string or vector of strings}

\item{extra_stops}{additional words to add to stopword list}
}
\value{
A ratio score between 0 and 100
}
\description{
Computes the Fuzzy Ratio Between Two Strings
}
\details{
Fuzzy ratio is the simplest function in the package. It simply returns the ratio of the number of unshared characters in two strings to the length of the two strings.
}
\examples{
cars$distsp <- paste(cars$dist, '.0', sep = '')
cars <- cbindr(cars, 'distsp', '.', c('orgdist', 'zero'))

}

