% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyOperation.R
\name{fuzzy.firing}
\alias{fuzzy.firing}
\title{Fuzzy rule firing}
\usage{
fuzzy.firing(operator, x.mf, ante.mf, lower, upper)
}
\arguments{
\item{operator}{t-norm operator}

\item{x.mf}{the fuzzy input membership function}

\item{ante.mf}{the antecedent membership function}

\item{lower}{lower bound of the input}

\item{upper}{upper bound of the input}
}
\value{
the rule firing strenth
}
\description{
To get the firing strength for the given input fuzzification membership function and the antecedent membership function in the domain of [lower, upper]
}
\examples{
x.mf <- x.fuzzification(gbell.fuzzification, 3, c(1,2))
ante.mf <- genmf(gbellmf, c(1,2,6))
firing.strength <- fuzzy.firing(min, x.mf, ante.mf, lower=0, upper=10)
firing.strength
}
\author{
Chao Chen
}
