% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyfun.R
\name{fuzzyfun}
\alias{fuzzyfun}
\title{Fuzzy version of a function}
\usage{
fuzzyfun(fun,
                fx,
                n = 1001,
                ...)
}
\arguments{
\item{fun}{A string with the function name.}

\item{fx}{A fuzzy number given by a \code{fzn} object.}

\item{n}{Number of subintervals for finding optimal values in the non-monotone case.}

\item{...}{Other arguments for function \code{fun}.}
}
\value{
A fuzzy number of class \code{fzn}.
}
\description{
It computes the fuzzy version of a function applied to a fuzzy number
\code{fx} given by a \code{fzn} object.
}
\examples{
fx <- fzn(c(0, 1), c(-1, 0), c(1, 0))
par(mfrow = c(2, 2))
plot(fx, main = "fx")
fxx <- fx * fx
plot(fxx, main = "fx * fx")
fx2 <- fuzzyfun("^", fx, y = 2) # It is not the same!
plot(fx2, main = "fx^2")
fz <- fuzzyfun("sqrt", fuzzyfun("abs", fx))
plot(fz, main = "sqrt(abs(fx))")

}
