% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyfy.R
\name{fuzzyfy}
\alias{fuzzyfy}
\title{Write a crisp number as a fuzzy number.}
\usage{
fuzzyfy(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A \code{fzn} object corresponding to the crisp number \code{x}.
}
\description{
This function creates a fuzzy number of class \code{fzn} from a crisp
numeric value, using as membership function the characteristic function of the number.
}
\examples{

}
