% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fuzzyr.match.fun}
\alias{fuzzyr.match.fun}
\title{fuzzyr.match.fun}
\usage{
fuzzyr.match.fun(FUN, descend = TRUE)
}
\arguments{
\item{FUN}{item to match as function: a function, symbol or character string.}

\item{descend}{logical; control whether to search past non-function objects.}
}
\description{
This is a modification of the original match.fun, where parent.frame(2) is changed to parent.env(environment()).
}
\details{
See \code{\link{match.fun}}.
}
