% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fzarray.R
\name{fzarray}
\alias{fzarray}
\title{Fuzzy array}
\usage{
fzarray(...)
}
\arguments{
\item{...}{Fuzzy numbers of class \code{fzn} or fuzzy arrays of class \code{fzarray}.
It can be a list of fuzzy numbers of class \code{fzn}.}
}
\value{
An object of class \code{fzarray}.
}
\description{
This function constructs a fuzzy array of class \code{fzarray} (and \code{fzn}), which is a list
of fuzzy numbers of class \code{fzn}. It can be used for adding elements to a fuzzy array,
or concatenate several fuzzy arrays.
}
\examples{
x <- fzn(alpha = c(0, 0.1, 0.5, 0.8, 1),
         l = c(1, 3, 3, 4.5, 5),
         u = c(12, 10, 9, 6.5, 6))
c <- fuzzyfy(1)
# Construct a fuzzy array with x and c
fza1 <- fzarray(x, c)
# Construct a fuzzy array from crisp values using function fuzzyfy
fza2 <- fuzzyfy(c(1, 2, 3))
# Construct a fuzzy array whose elements are x, c, 1, 2, 3 and x again
fza3 <- fzarray(fza1, fza2, x)

}
