% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fzunion.R
\name{fzunion}
\alias{fzunion}
\title{fzunion}
\usage{
fzunion(fx1,
               fx2)
}
\arguments{
\item{fx1}{A fuzzy number given by a \code{fzn} object.}

\item{fx2}{A fuzzy number given by a \code{fzn} object.}
}
\value{
A list with the two resulting fuzzy numbers.
}
\description{
This function adds alpha-cuts...
}
\examples{

}
