% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{gbellmf}
\alias{gbellmf}
\title{Generalised bell membership function}
\usage{
gbellmf(mf.params)
}
\arguments{
\item{mf.params}{The parameters c(a, b, c) for a generalised bell membership function}
}
\value{
The generalised bell membership function of x for a given pair of parameters,
where x is a generic element of U, which is the universe of discourse of a fuzzy set X
}
\description{
To specify a generalised bell membership function with a pair of particular parameters
}
\details{
This is not an external function. It should be used through \code{\link{genmf}}.
}
\examples{
mf <- gbellmf(c(1,2,3))
# This is the same as:
mf <- genmf('gbellmf', c(1,2,3))

evalmf(5, mf)
}
\author{
Chao Chen
}
