% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{genmf}
\alias{genmf}
\title{Fuzzy membership function generator}
\usage{
genmf(mf.type, mf.params)
}
\arguments{
\item{mf.type}{The membership function type}

\item{mf.params}{The parameters for a membership function}
}
\value{
The desired type of membership function f(x),
where x is a generic element of U, which is the universe of discourse for a fuzzy set
}
\description{
To generate the corresponding membership function f(x), also called fuzzy set,
according to type and parameters
}
\examples{
mf <- genmf(gbellmf, c(1,2,3))
evalmf(1:10, mf)
}
\author{
Chao Chen
}

