% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{gensurf}
\alias{gensurf}
\title{Produce a graphical evaluated fuzzy inference system.}
\usage{
gensurf(fis, ix1 = 1, ix2 = 2, ox1 = 1)
}
\arguments{
\item{fis}{A fis must be provided.}

\item{ix1}{Optional input (1)}

\item{ix2}{Optional input (2)}

\item{ox1}{Optional output}
}
\value{
A three dimensional graphical model generated from the fis and other optional parameters.
}
\description{
Produces a three dimensional graphical view of a specific fis object. This function is only  works for FIS structures with 3 variables. It will only work for 2 inputs, and 1 output.
}
\examples{
fis <- tipper()
gensurf(fis)
}
