% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.crisp.R
\name{is.crisp}
\alias{is.crisp}
\title{Crisp check.}
\usage{
is.crisp(fx)
}
\arguments{
\item{fx}{Any \code{fzn} object.}
}
\value{
Returns \code{TRUE} if \code{fx} is a crisp number and \code{FALSE} otherwise.
}
\description{
Checks whether an \code{fzn} object correspond to a crisp number or not.
}
