% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fzarray.R
\name{is.fzarray}
\alias{is.fzarray}
\title{Fuzzy array check}
\usage{
is.fzarray(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is an \code{fzarray} object (that is, it has "fzarray"
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks whether an R object is an \code{fzarray} or not.
}
