% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fzn.R
\name{is.fzn}
\alias{is.fzn}
\title{Fuzzy number check.}
\usage{
is.fzn(x)
}
\arguments{
\item{x}{Any \bold{R} object.}
}
\value{
Returns \code{TRUE} if its argument is a \code{fzn} number (that is, has "fzn"
  amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks whether an R object is of class \code{fzn} or not.
}
