% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{it2tipper}
\alias{it2tipper}
\title{Produces an example it2fis object for Waiter-Tipping.}
\usage{
it2tipper()
}
\value{
A fis object
}
\description{
A function used primarily for example purposes, it creates a it2 fis with two input (service & food), output variables (tip) and their membership functions.
}
\examples{
it2fis <- it2tipper()
}
