% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{linearmf}
\alias{linearmf}
\title{Linear membership function}
\usage{
linearmf(mf.params)
}
\arguments{
\item{mf.params}{The linear parameters, which is a vector of the size of input numbers plus 1}
}
\value{
A linear membership function
}
\description{
To specify a 1st order linear membership function with given parameters
}
\author{
Chao Chen
}
