% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{Degree of membership}
\usage{
membership(fx,
                  xout,
                  h = 0.01)
}
\arguments{
\item{fx}{A fuzzy number given by a \code{fzn} object.}

\item{xout}{A numeric vector of real values from which their degrees of membership will
be computed.}

\item{h}{Numerical, between 0 and 1. Step for alphas in "spline" interpolation.}
}
\value{
A list with \code{xout} and their corresponding degrees of membership.
}
\description{
This function computes the degree of membership of a vector of real values
in a given fuzzy number of class \code{fzn}.
}
\examples{

}
