% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_fzz.R
\name{piecewise_fzz}
\alias{piecewise_fzz}
\title{Construct a fuzzy number using a stepwise function}
\usage{
piecewise_fzz(extremos,
                 funciones,
                 valor)
}
\arguments{
\item{extremos}{A (ordered) list containing the extremes at which the different step functions are defined.}

\item{funciones}{A list containig of step functions, it is possible to use user functions defined in the usual form.}

\item{valor}{A number that the fuzzy stepwise-number is evaluated.}
}
\value{
A number, the value of the memebership function (between 0 and 1).
}
\description{
This function is used to specify a fuzzy number using a stepwise function.
}
\examples{
extrems<-c(12,15,18)
funciones<-c("exp","log")
val<-17
piecewise_fzz(extrem,funciones,val)

}
