% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fzn.R
\name{plot.fzn}
\alias{plot.fzn}
\title{Plot an fuzzy number}
\usage{
plot(fx,
            main = NULL,
            xlab = "x",
            ylab = "alpha",
            col = "black",
            plot_alpha = FALSE,
            h = 0.01)
}
\arguments{
\item{fx}{A fuzzy number defined by an object of \code{fzn} class.}

\item{main}{Argument passed to function plot.}

\item{xlab}{Argument passed to function plot.}

\item{ylab}{Argument passed to function plot.}

\item{col}{Argument passed to function plot.}

\item{plot_alpha}{Logical. If \code{TRUE}, it plots the alpha-cuts.}

\item{h}{Numerical, between 0 and 1. Step for alphas in "spline" graphic.}
}
\description{
Plots an fzn object.
}
\examples{
fx <- fzn(alpha = c(0, 0.1, 0.5, 0.8, 1),
         l = c(1, 3, 3, 4.5, 5),
         u = c(12, 10, 9, 6.5, 6),
         interp = "spline")
plot(fx, main = "Example of fuzzy number", plot_alpha = TRUE)
}
