% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{+.fzn}
\alias{+.fzn}
\title{Sum of fuzzy numbers}
\usage{
fx1 + fx2
}
\arguments{
\item{fx1}{First fuzzy (or crisp) number.}

\item{fx2}{Second fuzzy (or crisp) number.}
}
\value{
The fuzzy sum (fx1 + fx2) as an object of class \code{fzn}.
}
\description{
Sum two fuzzy numbers of class \code{fzn} or one fuzzy number to a crisp
number.
}
\examples{

}
