% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{readfis}
\alias{readfis}
\title{Read a fis object from a .fis file.}
\usage{
readfis(fileName)
}
\arguments{
\item{fileName}{Should be an absolute path given as a string to the file to be read, with escaped backslashes.}
}
\value{
A fis structure with its values generated from that of the files.
}
\description{
Reads a fis object from a file with the .fis extension, and converts it into a data structure to be used within the environment.
}
