% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{showfis}
\alias{showfis}
\title{Show a fis object.}
\usage{
showfis(fis)
}
\arguments{
\item{fis}{Requires a fis structure to be displayed.}
}
\value{
Returned the organised text regarding the fis is output to console.
}
\description{
Shows a fis and all its data in an ordered format on the console.
}
\examples{
fis <- tipper()
showfis(fis)
}
