% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{showGUI}
\alias{showGUI}
\title{Show a Graphic User Interface of fis object}
\usage{
showGUI(fis, advancedGUI = FALSE)
}
\arguments{
\item{fis}{Requires a fis structure to display a GUI.}

\item{advancedGUI}{TRUE/FALSE; if TRUE, an advanced GUI with more features is provided (provided by science@sboldt.com).}
}
\value{
Return the GUI to display membership function for input and output together with rules.
}
\description{
Show a Graphic User Interface to display membership function plots for input and output, rules and evaluate the fis.
}
\details{
This function is purposed to display all the membership plots and rules of fis object in Graphic User Interface (GUI). It also provide a function to evaluate the fis object.\cr

showGUI(fis) will display the GUI of fis object.
}
\examples{
fis <- tipper()
fis <- showGUI(fis)
}
\author{
Tajul Razak
}
