% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{showrule}
\alias{showrule}
\title{Showing rule from fis object}
\usage{
showrule(fis)
}
\arguments{
\item{fis}{A fis must be provided.}
}
\value{
Show the total of rules inside fis object
}
\description{
All the rule is showing from fis object
}
\examples{
fis <- tipper()
ruleList <- rbind(c(1,1,1,1,2), c(2,0,2,1,1), c(3,2,3,1,2))
fis <- addrule(fis, ruleList)
showrule(fis)
}
