% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a fuzzy number}
\usage{
simplify(fx,
                abstol_alpha = 1e-3,
                abstol_interp = 1e-6)
}
\arguments{
\item{fx}{A fuzzy  number given by a \code{fzn} object.}

\item{abstol_alpha}{Absolute tolerance for simplifying nearby alpha-cuts. If three different alpha-cuts
differ less than \code{abstol_alpha}, the middle one is eliminated. By default, it is 1e-3.}

\item{abstol_interp}{Absolute tolerance for simplifying taking into account interpolation. By default,
it is 1e-6.}
}
\value{
It returns a fuzzy number.
}
\description{
Simplify a fuzzy number given by an object of class \code{fzn}. It eliminates
unnecessary alpha-cuts.
}
\examples{
fx <- fzn(alpha = c(0, 0.2, 0.5, 0.8, 1),
          l = c(-1, 0, 1, 2, 3),
          u = c(8, 7, 6, 5, 4)) # interp = linear by default
par(mfrow = c(2, 2))
plot(fx, main = "fx (interp = linear)", plot_alpha = TRUE)
fx2 <- simplify(fx)
plot(fx2, main = "fx simplification using linear interpolation", plot_alpha = TRUE)
fx$interp <- "spline"
plot(fx, main = "fx (interp = spline)", plot_alpha = TRUE)
fx3 <- simplify(fx)
plot(fx3, main = "fx simplification using spline interpolation", plot_alpha = TRUE)
# For a sufficiently large abstol_alpha or abstol_interp, it returns a trapezoidal (or triangular) fuzzy number:
fx4 <- simplify(fx, abstol_alpha = 1)
plot(fx4, main = "Trapezoidal version of fx", plot_alpha = TRUE)

}
