% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{singleton.fuzzification}
\alias{singleton.fuzzification}
\title{Singleton Fuzzification}
\usage{
singleton.fuzzification(x, mf.params)
}
\arguments{
\item{x}{the crisp input}

\item{mf.params}{not used, singleton fuzzification does not need additional parameters}
}
\value{
The singleton MF at the crisp point x
}
\description{
To generate a fuzzy membership function based on singleton fuzzification for the given crisp input x
}
\examples{
mf <- singleton.fuzzification(3)
evalmf(1:10, mf)
}
\author{
Chao Chen
}

