% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyMF.R
\name{singletonmf}
\alias{singletonmf}
\title{Singleton membership function}
\usage{
singletonmf(mf.params)
}
\arguments{
\item{mf.params}{the particular singleton point}
}
\value{
The singleton membership function of x at the particular point,
where x is a generic element of U, which is the universe of discourse of a fuzzy set X
}
\description{
To specify a singleton membership function at the particular point
}
\details{
This is not an external function. It should be used through \code{\link{genmf}}.
}
\examples{
mf <- singletonmf(3)
# This is the same as:
mf <- genmf('singletonmf', 3)

evalmf(1:10, mf)
}
\author{
Chao Chen
}
