% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substring_ratio.R
\name{substr_ratio}
\alias{substr_ratio}
\title{Substring Ratio}
\usage{
substr_ratio(x1, x2, extra_stops = c())
}
\arguments{
\item{x1}{a single string or vector of strings}

\item{x2}{a single string or vector of strings}

\item{extra_stops}{additional words to add to stopword list}
}
\value{
A ratio score between 0 and 100
}
\description{
Computes Substring Ratio Between Two Strings
}
\details{
Substring Ratio returns the score of two strings based on the ratio of shared words to the length of both strings combined.
}
\examples{
cars$distsp <- paste(cars$dist, '.0', sep = '')
cars <- cbindr(cars, 'distsp', '.', c('orgdist', 'zero'))

}

