% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{*.fzarray}
\alias{*.fzarray}
\title{Product of fuzzy numbers vectors}
\usage{
fa1 * fa2
}
\arguments{
\item{fa1}{First fuzzy vector.}

\item{fa2}{Second fuzzy vector.}
}
\value{
The fuzzy product (fa1 * fa2) as an object of class \code{fzarray}.
}
\description{
Multiplies two arrays of fuzzy numbers of class \code{fzn} element-wise.
}
\examples{

}
