% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{tipper.ns}
\alias{tipper.ns}
\title{Produces an example non-singleton fis object for Waiter-Tipping.}
\usage{
tipper.ns()
}
\value{
A non-singleton fis object
}
\description{
A function used primarily for example purposes, it creates a nsfis with two input (service & food), output variables (tip) and their membership functions.
}
\examples{
fis <- tipper.ns()
}
\author{
Yu Zhao
}
