% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{tipper}
\alias{tipper}
\title{Produces an example fis object for Waiter-Tipping.}
\usage{
tipper()
}
\value{
A fis is return
}
\description{
A function used primarily for example purposes, it creates a fis with two input (service & food), output variables (tip) and their membership functions.
}
\examples{
fis <- tipper()
}
