% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{tipper.tsk}
\alias{tipper.tsk}
\title{Produces an example fis object (TSK type), which can also be optimised by ANFIS.}
\usage{
tipper.tsk()
}
\value{
A fis is return
}
\description{
A function used primarily for example purposes, it creates a fis with two input (service & food), output variables (tip) and their membership functions.
}
\examples{
fis <- tipper.tsk()
}
