% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{tipperGUI}
\alias{tipperGUI}
\title{Graphic User Interface for Waiter-Tipping}
\usage{
tipperGUI()
}
\value{
Return graphic user interface for Waiter-Tipping
}
\description{
Graphic User Interface for Waiter-Tipping to display the membership function (input & output) and rules.
}
\examples{
fis <- tipperGUI()
}
\author{
Tajul Razak
}
