% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{tipperGUI2}
\alias{tipperGUI2}
\title{Graphic User Interface for Waiter-Tipping (another style)}
\usage{
tipperGUI2()
}
\value{
Return graphic user interface for Waiter-Tipping
}
\description{
Another style of Graphic User Interface for Waiter-Tipping to display the membership function (input & output) and rules.
}
\examples{
fis <- tipperGUI2()
}
\author{
Tajul Razak
}
