% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.deadata_fzn.R
\name{to.deadata_fzn}
\alias{to.deadata_fzn}
\title{Transforms to deadata_fzn}
\usage{
to.deadata_fzn(x)
}
\arguments{
\item{x}{Any \code{deadata} or \code{deadata_fuzzy} object.}
}
\value{
Returns a \code{deadata_fzn} object.
}
\description{
Transforms a \code{deadata} or \code{deadata_fuzzy} object
into a \code{deadata_fzn} object.
}
