% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_set_ratio.R
\name{token_set_ratio}
\alias{token_set_ratio}
\title{Token Set Ratio}
\usage{
token_set_ratio(x1, x2, extra_stops = c())
}
\arguments{
\item{x1}{a single string or vector of strings}

\item{x2}{a single string or vector of strings}

\item{extra_stops}{additional words to add to stopword list}
}
\value{
A ratio score between 0 and 100
}
\description{
Computes the Token Set Ratio Between Two Strings
}
\details{
Both token_set_ratio and token_sort_ratio treat each word in a string as a "token." These tokens are compared across the two strings. token_set_ratio compares two strings, and returns the ratio of set of shared tokens (i.e. only the words appearing in *both* strings) to the length of the first string.
}
\examples{
cars$distsp <- paste(cars$dist, '.0', sep = '')
cars <- cbindr(cars, 'distsp', '.', c('orgdist', 'zero'))

}

