% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_sort_ratio.R
\name{token_sort_ratio}
\alias{token_sort_ratio}
\title{Token Sort Ratio}
\usage{
token_sort_ratio(x1, x2, extra_stops = c())
}
\arguments{
\item{x1}{a single string or vector of strings}

\item{x2}{a single string or vector of strings}

\item{extra_stops}{additional words to add to stopword list}
}
\value{
A ratio score between 0 and 100
}
\description{
Computes the Token Sort Ratio Between Two Strings
}
\details{
Both token_set_ratio and token_sort_ratio treat each word in a string as a "token." These tokens are compared across the two strings. token_sort_ratio does the same things as token_set_ratio, but sorts the tokens alphabetically first.
}
\examples{
cars$distsp <- paste(cars$dist, '.0', sep = '')
cars <- cbindr(cars, 'distsp', '.', c('orgdist', 'zero'))

}

