% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyInferenceSystem.R
\name{writefis}
\alias{writefis}
\title{Write a fis object to a .fis file.}
\usage{
writefis(fis, fileName = "fuzzy.fis")
}
\arguments{
\item{fis}{The fuzzy inference system data structure to be saved.}

\item{fileName}{filename}
}
\description{
Write a fis object to a file with the .fis extension.
}
