% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyfy.R
\name{fuzzyfy}
\alias{fuzzyfy}
\title{Write a crisp number as a fuzzy number}
\usage{
fuzzyfy(x)
}
\arguments{
\item{x}{A numeric value (or array or matrix), or a fuzzy number of class
\code{fzn}, or a fuzzy array of class \code{fzarray}. If \code{x} is of class
\code{fzn} or \code{fzarray} then the function \code{fuzzyfy} returns \code{x}.}
}
\value{
A \code{fzn} object corresponding to the crisp number x.
}
\description{
This function creates a fuzzy number of class \code{fzn} from a crisp
numeric value, using as membership function the characteristic function of the number.
Moreover, it can build an object of class \code{fzarray} from a numeric array.
}
\examples{
x <- fuzzyfy(3)
plotfzn(x)
xarray <- fuzzyfy(c(3, 4))

}
