% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.crisp.R
\name{is.crisp}
\alias{is.crisp}
\title{Crisp check}
\usage{
is.crisp(x)
}
\arguments{
\item{x}{Any \code{fzn} or \code{fzarray} object.}
}
\value{
Returns \code{TRUE} if \code{x} is a crisp number and \code{FALSE} otherwise.
If \code{x} is a fuzzy array, it returns a logical vector.
}
\description{
Checks whether an \code{fzn} object corresponds to a crisp number or not.
}
