% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a fuzzy number}
\usage{
simplify(x,
               abstol = 1e-6)
}
\arguments{
\item{x}{A fuzzy number of class \code{fzn} or a fuzzy array of class \code{fzarray}.}

\item{abstol}{Absolute tolerance for simplifying. By default, it is 1e-6.}
}
\value{
It returns a fuzzy number of class \code{fzn} or a fuzzy array of class \code{fzarray}.
}
\description{
Simplify a fuzzy number of class \code{fzn} or a fuzzy array of class \code{fzarray}.
It eliminates unnecessary alpha-cuts.
}
\examples{
x <- fzn(alpha = c(0, 0.2, 0.5, 0.8, 1),
         l = c(-1, 0, 1, 2, 3),
         u = c(8, 7, 6, 5, 4),
         interp = "approx")
x2 <- simplify(x)
xx <- fzn(alpha = c(0, 0.2, 0.5, 0.8, 1),
         l = c(-1, 0, 1, 2, 3),
         u = c(8, 7, 6, 5, 4),
         interp = "spline")
xx2 <- simplify(xx)
# For a sufficiently large abstol, it returns a trapezoidal (or triangular) fuzzy number:
x3 <- simplify(x, abstol = 0.25)

}
